feat: prepara código para rodar em qualquer PostgreSQL (local ou produção)
- CORS lê de CORS_ORIGIN (env var) com fallback localhost
- Web usa VITE_API_URL (env var) com fallback localhost
- dotenv condicional: só carrega .env se arquivo existir
- Extrai bootstrap-app.ts compartilhado (main.ts + vercel-handler.ts)
- Cria vercel-handler.ts para deploy serverless no Vercel
- Cria vercel.json para API (serverless) e Web (SPA)
- Adiciona script vercel-build com prisma generate
- Remove menção específica a Neon.tech no Swagger
- Atualiza .env.example com CORS_ORIGIN e VITE_API_URL
