+++ apps/api/.env.example
@@ -1,11 +1,25 @@
-# PostgreSQL (Local)
-DATABASE_URL="postgresql://postgres:postgres@localhost:5432/reserva_ai_db?schema=public"
-# Neon Database (Production/Staging)
-# DATABASE_URL="postgresql://user:password@endpoint.neon.tech/neondb?sslmode=require"
-# JWT Secret (generate with: openssl rand -base64 32)
-JWT_SECRET=your-super-secret-key-change-this
-# CORS - comma-separated origins allowed
-CORS_ORIGIN="https://<your-web-app>.vercel.app"
-# Port
-PORT=3000
+# ============================================
+# 🏠 Reserva Aí! - Configuração de Ambiente
+# ============================================
+# Copie este arquivo para .env e ajuste os valores
+# Funciona com QUALQUER PostgreSQL (local, Neon, Supabase, Railway, etc.)
+# ============================================
 
+# --- Banco de Dados ---
+# Local (Docker):
+DATABASE_URL="postgresql://postgres:postgres@localhost:5432/reserva_ai_db?schema=public"
+# Produção (Neon, Supabase, Railway - basta trocar a URL):
+# DATABASE_URL="postgresql://user:password@host:5432/dbname?sslmode=require"
+
+# --- JWT ---
+# Gere com: openssl rand -base64 32
+JWT_SECRET=your-super-secret-key-change-this
+
+# --- CORS ---
+# Local (desenvolvimento):
+CORS_ORIGIN="http://localhost:5173,http://localhost:5174"
+# Produção (seu domínio do Vercel):
+# CORS_ORIGIN="https://seu-app.vercel.app"
+
+# --- Porta ---
+PORT=3000
