+++ apps/web/.env.example
@@ -1,0 +1,5 @@
+# URL base da API
+# Local:
+VITE_API_URL="http://localhost:3000/api/v1"
+# Produção (substitua pela URL do seu backend no Vercel):
+# VITE_API_URL="https://seu-api-app.vercel.app/api/v1"
