+++ apps/web/src/api/http.ts
@@ -3,7 +3,7 @@ import { router } from "@/router";
 import { useLoading } from "@/modules/shared/composables/useLoading";
 
 export const http = axios.create({
-    baseURL: "http://localhost:3000/api/v1",
+    baseURL: import.meta.env.VITE_API_URL || "http://localhost:3000/api/v1",
 });
 
 const { start, finish } = useLoading();
